Add DBAL's MasterSlaveConnection support to mysql-come-back
