+++ src/ConnectionInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL;
+
+use Doctrine\DBAL\Driver\Connection as DriverConnection;
+
+interface ConnectionInterface extends DriverConnection
+{
+}
